[Native][tests] Set up library paths for 1st and 2nd stages separately
Set up library paths for 1st and 2nd stages of Kotlin/Native compilation
separately (by analogy with Kotlin/JS and Kotlin/Wasm test pipelines):
* `NativeFirstStageEnvironmentConfigurator` - use only direct
  dependencies between test modules (i.e. no transitive deps)
* `NativeSecondStageEnvironmentConfigurator` - use full dependencies
  between test modules (including transitive dependencies).

^KT-61096
