+++ compiler/tests-common-new/testFixtures/org/jetbrains/kotlin/test/services/configuration/NativeFirstStageEnvironmentConfigurator.kt
@@ -7,6 +7,7 @@ package org.jetbrains.kotlin.test.services.configuration
 
 import org.jetbrains.kotlin.config.CompilerConfiguration
 import org.jetbrains.kotlin.konan.config.*
+import org.jetbrains.kotlin.konan.library.KlibNativeDistributionLibraryProvider
 import org.jetbrains.kotlin.konan.target.CompilerOutputKind
 import org.jetbrains.kotlin.test.model.TestModule
 import org.jetbrains.kotlin.test.services.CompilationStage
@@ -23,6 +24,18 @@ class NativeFirstStageEnvironmentConfigurator(
 
     override fun configureCompilerConfiguration(configuration: CompilerConfiguration, module: TestModule) {
         super.configureCompilerConfiguration(configuration, module)
+
+        // Load only direct dependencies (i.e. no transitive dependencies).
+        val dependencies = module.regularDependencies.map { getKlibArtifactDir(testServices, it.dependencyModule.name).absolutePath }
+        val friends = module.friendDependencies.map { getKlibArtifactDir(testServices, it.dependencyModule.name).absolutePath }
+
+        val runtimeDependencies = getRuntimeLibraryProviders(module).flatMap { provider ->
+            // Ignore `KlibNativeDistributionLibraryProvider`, because it is anyway applied in loadNativeKlibs().
+            if (provider is KlibNativeDistributionLibraryProvider) emptyList() else provider.getLibraryPaths()
+        }
+
+        configuration.konanLibraries = runtimeDependencies + dependencies + friends
+        configuration.konanFriendLibraries = friends
 
         configuration.konanProducedArtifactKind = CompilerOutputKind.LIBRARY
         configuration.konanOutputPath = getKlibArtifactDir(testServices, module.name).absolutePath
